requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.35
pdfplumber==0.10.3
pytesseract>=0.3.13
Pillow>=10.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
cryptography==41.0.7

